requirements.txt
beatifulsoup4
pytest
python-dotenv
pyqt
requests
